[NFC] Split objcImpl cdecl tests from objc tests
We need to run the cdecl tests on all platforms, not just ones that support ObjC interop.

This commit simply increases test coverage of things that fortunately already worked.
