+++ test/IRGen/Inputs/objc_implementation.h
@@ -24,10 +24,6 @@
 
 @end
 
-extern void implFunc(int param);
-extern void implFuncCName(int param) __asm__("_implFuncAsmName");
-
-
 @interface NoImplClass
 
 - (void)noImplMethod:(int)param;
