+++ test/IRGen/objc_implementation.m
@@ -27,11 +27,3 @@ int printf(const char * restrict format, ...);
 }
 
 @end
-
-void implFunc(int param) {
-  printf("implFunc");
-}
-
-void implFuncCName(int param) {
-  printf("implFuncCName");
-}
