+++ test/Interpreter/Inputs/cdecl_implementation.h
@@ -1,0 +1,1 @@
+extern void implFunc(int param);
