+++ test/Interpreter/Inputs/cdecl_implementation.modulemap
@@ -1,0 +1,3 @@
+module cdecl_implementation {
+  header "cdecl_implementation.h"
+}
