+++ test/Interpreter/cdecl_implementation.swift
@@ -1,0 +1,12 @@
+// RUN: %target-run-simple-swift(-import-objc-header %S/Inputs/cdecl_implementation.h -D TOP_LEVEL_CODE -swift-version 5) %s | %FileCheck %s
+// REQUIRES: executable_test
+
+@_objcImplementation @_cdecl("implFunc") public func implFunc(_ param: Int32) {
+  print("implFunc(\(param))")
+}
+
+// `#if swift` to ignore the inactive branch's contents
+#if swift(>=5.0) && TOP_LEVEL_CODE
+implFunc(2023 - 34)
+// CHECK: implFunc(1989)
+#endif
