+++ test/lit.cfg
@@ -2348,6 +2348,10 @@ elif not kIsWindows and not run_os == 'wasi':
         lit_config.note('Testing with the just-built libraries')
 
     lit_config.note('Library load path: {0}'.format(os.path.pathsep.join(target_stdlib_path)))
+    config.target_run_additional_libs = (
+        "/usr/bin/env " +
+        construct_library_path_env(target_stdlib_path + [r'\1']) +
+        config.target_run)
     config.target_run = (
         "/usr/bin/env " +
         construct_library_path_env(target_stdlib_path) +
@@ -2450,6 +2454,9 @@ if not getattr(config, 'target_run_simple_swift', None):
         '%%line-directive %%t/main.swift -- '
         '%s %%t/a.out'
         % (config.target_build_swift, mcp_opt, config.target_codesign, config.target_run))
+
+if not getattr(config, 'target_run_additional_libs', None):
+    config.target_run_additional_libs = config.target_run
 
 # FIXME: why can we not use %rth and have that be expanded out?
 config.target_resilience_test = (
@@ -2567,6 +2574,8 @@ config.substitutions.append(('%target-run-simple-swift', config.target_run_simpl
 config.substitutions.append(('%target-run-stdlib-swiftgyb', config.target_run_stdlib_swiftgyb))
 config.substitutions.append(('%target-run-stdlib-swift', config.target_run_stdlib_swift))
 config.substitutions.append(('%target-repl-run-simple-swift', subst_target_repl_run_simple_swift))
+config.substitutions.append(('%target-run-additional-libs\(([^)]+)\)',
+                             SubstituteCaptures(config.target_run_additional_libs)))
 config.substitutions.append(('%target-run', config.target_run))
 config.substitutions.append(('%target-jit-run', subst_target_jit_run))
 config.substitutions.append(('%target-build-swift-dylib\(([^)]+)\)', config.target_build_swift_dylib))
